refactor: users 변수명을 명확한 의미의 invitees 로 변경
why:
- 초대된 상황에 맞추어 사용자라는 변수명을 invitee 로 변경하여 코드 가독성을 향상시킨다
